requirements.txt
asgiref
attr
backports.zoneinfo
Django
django-debug-toolbar
mysqlclient
sqlparse
djangorestframework
matplotlib
packaging
plotly
tenacity
remote-pdb